python binding: workaround for python 3.5 distutils bug
The cffi compile method uses python distutils under the hood,
so it is subject to bugs in distutils. Turns out there is a
bug in python 3.5 distutils for unix platforms: absolute paths
for library dirs do not work, causing link failures.

Avoid this bug, when present, by ensuring path to couchbase build dir
is always relative.