Fix Nim bindings on Linux, and updated README
This adds a conditional to select the correct name for the dynamic
library based on platform. It has a switch for Windows as well, but this
remains untested. Also updated the README to include how to run on Linux
and explain why `nimble build` doesn't do anything (because it's a
library package and not a binary package, `nimble install` would put the
library in your nimble path so you could import it directly from your
projects).