BuildPyCBL.py: support cross-compile for python bindings
Introduce optional --python_includedir argument to BuildPyCBL.py,
so that the target's headers (e.g. from an SDK) can be supplied instead
of using python headers from the build system itself.

It is always a good idea to use the right set of headers, but python
headers do a sanity check that fails the build if the target is 32-bit
but headers come from a 64-bit build system:

  #error "LONG_BIT definition appears wrong for platform (bad gcc/glibc config?)."