Fix various style issues
This fixes various style issues and brings the code closer in line with
the NEP-1 style guide. All the files where passed through the nimpretty
tool with indents set to 2 spaces (which is the NEP-1 specified amount)
and then manually corrected comments to linewrap to 80 characters for
most comments. Also fixed some comment blocks here and there that where
normal comments and not documentation comments.