Fix various style issues in the Nim bindings (#58)
* Fix Nim bindings on Linux, and updated README

This adds a conditional to select the correct name for the dynamic
library based on platform. It has a switch for Windows as well, but this
remains untested. Also updated the README to include how to run on Linux
and explain why `nimble build` doesn't do anything (because it's a
library package and not a binary package, `nimble install` would put the
library in your nimble path so you could import it directly from your
projects).

* Fix Windows DLL name

* Remove some dead code

* Fix various style issues

This fixes various style issues and brings the code closer in line with
the NEP-1 style guide. All the files where passed through the nimpretty
tool with indents set to 2 spaces (which is the NEP-1 specified amount)
and then manually corrected comments to linewrap to 80 characters for
most comments. Also fixed some comment blocks here and there that where
normal comments and not documentation comments.