Fixed Android build by updating CBLDatabaseChange
The latest change to LiteCore added a field to C4CollectionObserver::Change,
and CBLDatabaseChange needs to remain identical to that struct, so I've added
a corresponding field.
For some reason having to do with struct alignment this did not break the build
on macOS, so I didn't notice it right away. It did break Android though.