Updated Doxyfile for newer Doxygen, added Xcode scheme
Doxygen warned that the config file was out of date, so I allowed it to update
the file. Looks like this bumped the version number at the top, updated a bunch
of comments, added new configs with default values, etc.

I did not make manual changes to the file.

I also added an Xcode scheme to run the "Doxygen" target, which builds the docs.