CBL-7553 : Use the actual winner and loser revision id when calling c4doc_resolveConflict
Previously, prior version vector, all platforms always passed the remote revision ID as the winner when resolving conflicts to allow the resolved rev to avoid futher conflict when pusing the rev back to the server.

With version vector documents, LiteCore's resolveConflict() now requires the actual winner and loser revision IDs - the logic to generate the resolved rev or version will be done in resolveConflict() based on the type of the rev (rev-tree or version). Updated the conflict resolution logic to pass the required values.